Transactions removed

+++ persistence/src/main/java/cz/muni/fi/pa165/persistence/DAO/LeaseDAOImpl.java
@@ -6,7 +6,7 @@ import cz.muni.fi.pa165.persistence.Entities.Person;
 import java.util.Date;
 import java.util.List;
 import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
+import javax.persistence.PersistenceContext;
 
 /**
  * Implementation of LeaseDAO interface Implements functionality of lease
@@ -16,39 +16,38 @@ import javax.persistence.EntityManagerFactory;
  */
 public class LeaseDAOImpl implements LeaseDAO {
 
-    private EntityManagerFactory emf;
+    @PersistenceContext(name="carManagementSystem-unit")
+    private EntityManager em;
 
-    public LeaseDAOImpl(EntityManagerFactory emf) {
-        this.emf = emf;
+    public LeaseDAOImpl(EntityManager entityMf) {
+        if (entityMf == null) throw new IllegalArgumentException("the argument must be set");
+
+        em = entityMf;
     }
 
-    public EntityManagerFactory getEntityManagerFactory() {
-        return emf;
+    public EntityManager getEntityManagerFactory() {
+        return em;
     }
 
     @Override
     public void createLease(Lease lease) {
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-        em.persist(lease);
-        em.getTransaction().commit();
-        em.close();
+        if (lease == null) throw new IllegalArgumentException("Lease is null");
+ 
+        em.persist(lease); 
     }
 
     @Override
-    public void updateLease(Lease lease, int leaseId) {
-        EntityManager em = emf.createEntityManager();
+    public void updateLease(Lease lease, int leaseId) {  
+        if (lease == null) throw new IllegalArgumentException("lease is null");
+        if (leaseId < 0 ) throw new IllegalArgumentException("Wrong input for id");
         
-
         int carMileage = lease.getCarMileage();
         Date dateOfLease = lease.getDateOfLease();
         Date dateOfReturn = lease.getDateOfReturn();
         Boolean isClosed = lease.getIsClosed();
         Car car = lease.getCar();
         Person person = lease.getPerson();
-        
-        em.getTransaction().begin();
-        
+           
         Lease lease1 = (Lease)em.find(Lease.class ,leaseId);
 
         lease1.setCarMileage(carMileage);
@@ -57,47 +56,30 @@ public class LeaseDAOImpl implements LeaseDAO {
         lease1.setIsClosed(isClosed);
         lease1.setCar(car);
         lease1.setPerson(person);
-
-        em.getTransaction().commit();
-        em.close();
     }
 
     @Override
     public void deleteLease(int leaseId) {
-
-        EntityManager em = emf.createEntityManager();
-        em.getTransaction().begin();
-
-        Lease lease = (Lease)em.find(Lease.class ,leaseId);
+        if (leaseId < 0) throw new IllegalArgumentException("Wrong input for id");
         
+        Lease lease = (Lease)em.find(Lease.class ,leaseId);  
         em.remove(lease);
-        
-        em.getTransaction().commit();
-        em.close();
     }
     
     @Override
     public List getLeasesByPerson(Person person) {
-
-        EntityManager em = emf.createEntityManager();
+        if (person == null) throw new IllegalArgumentException("person is null");
         
-        em.getTransaction().begin();
-
         String query = "SELECT k FROM Lease k WHERE k.person= :person";//TODO
         List<Lease> leases = em.createQuery(query).setParameter("person", person).getResultList();
-
-        em.getTransaction().commit();
-        em.close();
 
         return leases;
     }
 
     @Override
     public List getAllLeases(Date from, Date until) {
-
-        EntityManager em = emf.createEntityManager();
-        
-        em.getTransaction().begin();
+        if (from == null) throw new IllegalArgumentException("Wrong input for date from");
+        if (until == null) throw new IllegalArgumentException("Wrong input for date until");
 
         String query = "SELECT k FROM Lease k WHERE k.dateOfLease BETWEEN :from AND :until";
         
@@ -105,30 +87,16 @@ public class LeaseDAOImpl implements LeaseDAO {
                 setParameter("from", from).setParameter("until", until)
                 .getResultList();
 
-        em.getTransaction().commit();
-        em.close();
-
         return leases;
     }
     
 
     @Override
     public Lease getLeaseByID(Integer ID) {
-         //create Entity Manager
-        //EntityManagerFactory emf = Persistence.createEntityManagerFactory("carManagementSystem-unit");
-        EntityManager em = emf.createEntityManager();
-
-        //begin of a transaction
-        em.getTransaction().begin();
-
         //actual query
         String sql = "SELECT l FROM Lease l WHERE l.id=:leaseID";
         Lease lease = em.createQuery(sql, Lease.class)
                 .setParameter("leaseID", ID).getResultList().get(0);
-
-        //commiting changes and closing entity manager
-        em.getTransaction().commit();
-        em.close();
 
         return lease;
     }
